converter/pyproject.toml
[tool.poetry]
name = "converter"
version = "0.1.0"
description = ""
authors = ["coryphoenixxx <coryphoenixxx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
webptools = "^0.0.9"
bottle = "^0.12.25"
python-decouple = "^3.8"
gunicorn = "^21.2.0"
python-magic = "^0.4.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
